# chore(designer): Removed a lot of unused CSS rules