--- a/libs/designer-ui/src/lib/arrayeditor/arrayeditor.less
+++ b/libs/designer-ui/src/lib/arrayeditor/arrayeditor.less
@@ -28,10 +28,6 @@
         box-sizing: border-box;
         width: calc(100% - @command-icon-width - @parameter-label-inputbox-margin);
 
-        .msla-input-parameter-box {
-          width: calc(100% - @parameter-label-inputbox-margin);
-        }
-
         .msla-input-parameter-error {
           margin-left: 0;
         }