--- a/libs/designer-ui/src/lib/card/config/config.less
+++ b/libs/designer-ui/src/lib/card/config/config.less
@@ -18,10 +18,6 @@
   }
 }
 
-.msla-card-config-button-container {
-  text-align: center;
-}
-
 .msla-card-config-item {
   margin: 6px @parameter-label-inputbox-margin;
   display: flex;
@@ -79,39 +75,6 @@
   }
 }
 
-.msla-dropdown-icon {
-  display: inline-block;
-  position: relative;
-  margin-right: 8px;
-}
-
-.msla-dropdown-option {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.msla-card-config-table-item {
-  display: block;
-
-  label {
-    display: inline;
-  }
-}
-
-.msla-card-config-button-container {
-  margin-top: 10px;
-  margin-right: 10px;
-
-  .msla-card-button-primary,
-  .msla-card-button-secondary {
-    margin-top: 10px;
-    margin-right: 10px;
-    min-width: 150px;
-  }
-}
-
 .config-item-narrow-width-mixin() {
   .msla-card-config-item {
     > label {
@@ -135,10 +98,6 @@
 }
 
 @media only screen and (max-width: @designer-desktop-max-width) {
-  .config-item-narrow-width-mixin();
-}
-
-.msla-connectionwizard-canvas--narrow {
   .config-item-narrow-width-mixin();
 }