--- a/libs/designer-ui/src/lib/card/function/function.less
+++ b/libs/designer-ui/src/lib/card/function/function.less
@@ -4,12 +4,6 @@
 .msla-card-editor {
   text-align: left;
 
-  .editor-header {
-    .msla-input-parameter-box {
-      border: none;
-    }
-  }
-
   .editor-body {
     .msla-monaco {
       border: 1px solid @defaultBorderColor;