--- a/libs/designer-ui/src/lib/card/swaggerendpoint/swaggerendpoint.less
+++ b/libs/designer-ui/src/lib/card/swaggerendpoint/swaggerendpoint.less
@@ -1,7 +1,0 @@
-.msla-swagger-endpoint-box {
-  .placeholder();
-}
-
-.msla-card-swagger-body.msla-card-config-item {
-  display: block;
-}