--- a/libs/designer-ui/src/lib/common.less
+++ b/libs/designer-ui/src/lib/common.less
@@ -32,17 +32,6 @@
   visibility: hidden;
 }
 
-.msla-padded-text {
-  font-family: @font-family;
-  font-size: @card-body-font-size;
-  padding: 8px;
-}
-
-.msla-two-columns {
-  display: inline-block;
-  width: 50%;
-}
-
 .msla-theme-dark {
   .msla-button {
     color: @brandColorLight;