--- a/libs/designer-ui/src/lib/dictionary/dictionaryeditor.less
+++ b/libs/designer-ui/src/lib/dictionary/dictionaryeditor.less
@@ -29,10 +29,6 @@
         box-sizing: border-box;
         width: calc(100% - @command-icon-width);
 
-        .msla-input-parameter-box {
-          width: calc(100% - @parameter-label-inputbox-margin);
-        }
-
         .msla-input-parameter-error {
           margin-left: 0;
         }