--- a/libs/designer-ui/src/lib/fabric.less
+++ b/libs/designer-ui/src/lib/fabric.less
@@ -1,19 +1,5 @@
 @import (reference) './variables.less';
 
-.msla-modal-dialog {
-  div[class^='ms-Dialog-main'] {
-    width: 80vw;
-
-    @media (min-width: 480px) {
-      max-width: none;
-    }
-
-    @media only screen and (max-width: 800px) {
-      width: 100%;
-    }
-  }
-}
-
 /* Checkbox (ui/checkbox) */
 .msla-checkbox {
   align-items: baseline;