--- a/libs/designer-ui/src/lib/flyout/flyout.less
+++ b/libs/designer-ui/src/lib/flyout/flyout.less
@@ -13,18 +13,6 @@
   }
 }
 
-.msla-combobox {
-  .msla-input-control-token-expression {
-    .msla-flyout {
-      .msla-flyout-icon {
-        display: inline-block;
-        height: 15px;
-        width: 15px;
-      }
-    }
-  }
-}
-
 .msla-flyout-callout {
   div[role='dialog'] {
     padding: 8px;