--- a/libs/designer-ui/src/lib/index.ts
+++ b/libs/designer-ui/src/lib/index.ts
@@ -35,7 +35,6 @@ export * from './editor/base';
 export * from './editor';
 export * from './editor/monaco';
 export * from './editor/string';
-export * from './errorboundary';
 export * from './errorsPanel';
 export * from './eventhandler';
 // export * from './featurefeedback';