--- a/libs/designer-ui/src/lib/settings/settings.less
+++ b/libs/designer-ui/src/lib/settings/settings.less
@@ -163,38 +163,3 @@
     margin-top: 9px;
   }
 }
-
-.msla-operation-setting-section {
-  text-align: left;
-
-  .msla-operation-setting {
-    display: flex;
-
-    .msla-setting-label {
-      margin-top: 5px;
-      width: @parameter-label-width;
-      flex-shrink: 0;
-
-      label {
-        display: inline-block;
-      }
-
-      .msla-flyout-icon {
-        width: 14px;
-        height: 14px;
-        margin: 0;
-      }
-    }
-  }
-
-  .ms-Toggle-label {
-    display: inline-block;
-    width: @parameter-label-width;
-    font-size: @card-label-font-size;
-    vertical-align: middle;
-  }
-
-  [class^='ms-Toggle-slider'] {
-    display: inline-block;
-  }
-}