… Localize/lang/strings.json
… libs/designer-ui/src/lib/card/card.less
… libs/designer-ui/src/lib/card/cardmonitor.less
… libs/designer-ui/src/lib/card/operationconfigurations/operationconfigurations.less
… libs/designer-ui/src/lib/card/query/query.less
… libs/designer-ui/src/lib/conditioncontrol/conditioncontrol-fields.less
… libs/designer-ui/src/lib/conditioncontrol/conditioncontrol.less
… libs/designer-ui/src/lib/connectioncontainer.less
… libs/designer-ui/src/lib/errorboundary/__test__/__snapshots__/errorboundary.spec.tsx.snap
… libs/designer-ui/src/lib/errorboundary/__test__/errorboundary.spec.tsx
… libs/designer-ui/src/lib/errorboundary/index.tsx
… libs/designer-ui/src/lib/inputcontrol/inputcontrol.less
… libs/designer-ui/src/lib/itemspicker/itemspicker.less
… libs/designer-ui/src/lib/pager/pager.less
… libs/designer-ui/src/lib/panel/panel.less
… libs/designer-ui/src/lib/recommendation/recommendation.less
… libs/designer-ui/src/lib/rulescontrol/rulescontrol.less
… libs/designer-ui/src/lib/schemadataeditorcard.less
… libs/designer-ui/src/lib/templates.less