## Main Changes

Removed a lot of unused CSS rules.

I noticed when investigating perf that we had a **lot** of rules that were being computed, but didn't actually have any intended elements in our project and would be rerun with no benefit.  
Removing these did decrease CSS redraw compute time in my test workflow from ~6ms to ~2ms (I know a staggering 4ms improvement)
This does happen on _every_ draw though so it's a 4ms that happens very often.